FossilOrigin-Name: a54be6e041a9185787a22b86603dcb0654e5e4af71225b556d1b6279f8520ad8

+++ ext/lsm1/lsm-test/lsmtest_tdb.c
@@ -767,7 +767,8 @@ int tdb_open(const char *zLib, const char *zDb, int bClear, TestDb **ppDb){
   if( *zSpec=='\0' ) zSpec = 0;
 
   for(i=0; i<ArraySize(aLib); i++){
-    if( strlen(aLib[i].zName)==nLib && 0==memcmp(zLib, aLib[i].zName, nLib) ){
+    if( (int)strlen(aLib[i].zName)==nLib
+        && 0==memcmp(zLib, aLib[i].zName, nLib) ){
       rc = aLib[i].xOpen(zSpec, (zDb ? zDb : aLib[i].zDefaultDb), bClear, ppDb);
       if( rc==0 ){
         (*ppDb)->zLibrary = aLib[i].zName;
